feat: 更新 EmbyAPI 兼容性报告与后端逻辑
- 在 `RemoteImageResult` 中新增 `StartIndex` 字段，支持分页请求。
- 修改失败重试逻辑，明确至多 3 次 HTTP 尝试，提升稳定性。
- 更新 `UserPolicyDto` 中 `max_active_sessions` 字段的序列化，确保与 Emby/Sakura 兼容。
- 增加 SQL 索引以优化查询性能，支持快速查找。
- 改进媒体项 DTO 转换逻辑，提升代码可读性与性能。
- 统一处理远端播放信息缓存，确保在失效时能重新拉取数据。
- 优化并发下载逻辑，提升媒体下载效率与稳定性。

+++ backend/src/models.rs
@@ -506,7 +506,7 @@ pub struct UserPolicyDto {
     pub enable_user_preference_access: bool,
     pub max_parental_rating: Option<i32>,
     pub max_parental_sub_rating: Option<i32>,
-    #[serde(rename = "SimultaneousStreamLimit")]
+    #[serde(rename = "SimultaneousStreamLimit", alias = "MaxActiveSessions")]
     pub max_active_sessions: i32,
     pub invalid_login_attempt_count: i32,
     pub login_attempts_before_lockout: i32,
@@ -1639,7 +1639,7 @@ pub struct TranscodingInfoDto {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub bitrate: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub audio_bitrate: Option<i32>,
+    pub audio_bitrate: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub video_bitrate: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
